Split: TestProeinDigestion
