+++ mzLib/Test/Proteomics/ProteolyticDigestion/ProteinDigestionTests.cs
@@ -1,4 +1,4 @@
-﻿using Chemistry;
+using Chemistry;
 using MassSpectrometry;
 using MzLibUtil;
 using NUnit.Framework;
@@ -26,7 +26,7 @@ namespace Test
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    public static class TestProteinDigestion
+    public static class ProteinDigestionTests
     {
         private static Stopwatch Stopwatch { get; set; }
 
@@ -41,21 +41,6 @@ namespace Test
         public static void TearDown()
         {
             Console.WriteLine($"Analysis time: {Stopwatch.Elapsed.Hours}h {Stopwatch.Elapsed.Minutes}m {Stopwatch.Elapsed.Seconds}s");
-        }
-
-        [Test]
-        public static void ProteaseLoader()
-        {
-            string path1 = Path.Combine(TestContext.CurrentContext.TestDirectory, "ProteaseFilesForLoadingTests", "TestProteases_badMod.tsv");
-            string path2 = Path.Combine(TestContext.CurrentContext.TestDirectory, "ProteaseFilesForLoadingTests", "TestProteases_badMod_dupName.tsv");
-            string path3 = Path.Combine(TestContext.CurrentContext.TestDirectory, "ProteaseFilesForLoadingTests", "TestProteases_dupName.tsv");
-            string path4 = Path.Combine(TestContext.CurrentContext.TestDirectory, "ProteaseFilesForLoadingTests", "TestProteases_Mod_dupName.tsv");
-            var proteaseMods = PtmListLoader.ReadModsFromFile(Path.Combine(TestContext.CurrentContext.TestDirectory, "ModificationTests", "ProteaseMods.txt"), out var errors).ToList();
-            
-            NUnit.Framework.Assert.Throws<MzLibUtil.MzLibException>(() => ProteaseDictionary.LoadProteaseDictionary(path1, proteaseMods)); 
-            NUnit.Framework.Assert.Throws<MzLibUtil.MzLibException>(() => ProteaseDictionary.LoadProteaseDictionary(path2, proteaseMods));
-            NUnit.Framework.Assert.Throws<MzLibUtil.MzLibException>(() => ProteaseDictionary.LoadProteaseDictionary(path3, proteaseMods));
-            NUnit.Framework.Assert.Throws<MzLibUtil.MzLibException>(() => ProteaseDictionary.LoadProteaseDictionary(path4, proteaseMods));
         }
 
         [Test]
@@ -74,7 +59,7 @@ namespace Test
                 protease: protease1.Name,
                 maxMissedCleavages: 0,
                 minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);            
+                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
             List<Modification> variableModifications1 = new List<Modification>();
 
             var protease2 = proteaseDict["CNBr_old"];
@@ -94,7 +79,7 @@ namespace Test
             List<Modification> variableModifications3 = new List<Modification>();
 
             var peps1 = prot.Digest(digestionParams1, new List<Modification>(), variableModifications1).ToList();
-            var peps2 = prot.Digest(digestionParams2, new List<Modification>(), variableModifications2).ToList(); 
+            var peps2 = prot.Digest(digestionParams2, new List<Modification>(), variableModifications2).ToList();
             var peps3 = prot2.Digest(digestionParams3, new List<Modification>(), variableModifications1).ToList();
 
             Assert.AreNotEqual(null, protease3.CleavageMod);
@@ -118,166 +103,7 @@ namespace Test
             Assert.AreEqual(882.39707781799996, peps1[0].MonoisotopicMass);
             Assert.AreEqual(930.400449121, peps1[1].MonoisotopicMass);
         }
-        /// <summary>
-        /// Tests that the embedded proteases.tsv resource exists, has the correct naming pattern,
-        /// and can be loaded successfully with expected protease definitions.
-        /// </summary>
-        [Test]
-        public static void LoadProteaseDictionary_EmbeddedResource_ExistsAndLoads()
-        {
-            // Verify the embedded resource exists with the expected naming pattern
-            var assembly = Assembly.GetAssembly(typeof(ProteaseDictionary));
-            var resourceNames = assembly.GetManifestResourceNames();
-            Assert.That(resourceNames, Contains.Item("Proteomics.ProteolyticDigestion.proteases.tsv"),
-                $"Expected embedded resource not found. Available resources: {string.Join(", ", resourceNames)}");
 
-            // Verify it loads successfully and contains expected proteases via ResetToDefaults
-            ProteaseDictionary.ResetToDefaults();
-            var dictionary = ProteaseDictionary.Dictionary;
-            Assert.That(dictionary, Is.Not.Null);
-            Assert.That(dictionary.Count, Is.GreaterThan(0));
-
-            // Verify well-known proteases exist with expected properties
-            Assert.That(dictionary.ContainsKey("trypsin|P"), Is.True);
-            Assert.That(dictionary["trypsin|P"].CleavageSpecificity, Is.EqualTo(CleavageSpecificity.Full));
-            Assert.That(dictionary["trypsin|P"].DigestionMotifs.Count, Is.EqualTo(2)); // K[P]| and R[P]|
-        }
-
-        /// <summary>
-        /// Tests backward compatibility for old-style protease names.
-        /// Names like "chymotrypsin (don't cleave before proline)" should automatically
-        /// resolve to "chymotrypsin|P".
-        /// </summary>
-        [Test]
-        public static void GetProtease_OldStyleName_ResolvesToNewFormat()
-        {
-            // Reset to defaults to ensure clean state
-            ProteaseDictionary.ResetToDefaults();
-
-            // Test various old-style naming patterns
-            var testCases = new[]
-            {
-                ("chymotrypsin (don't cleave before proline)", "chymotrypsin|P"),
-                ("trypsin (don't cleave before proline)", "trypsin|P"),
-                ("Lys-C (don't cleave before proline)", "Lys-C|P"),
-            };
-
-            foreach (var (oldName, expectedNewName) in testCases)
-            {
-                // Verify normalization
-                string normalizedName = ProteaseDictionary.NormalizeProteaseName(oldName);
-                Assert.That(normalizedName, Is.EqualTo(expectedNewName), 
-                    $"Failed to normalize '{oldName}' to '{expectedNewName}'");
-
-                // Verify GetProtease works with old name
-                var protease = ProteaseDictionary.GetProtease(oldName);
-                Assert.That(protease, Is.Not.Null, $"GetProtease failed for '{oldName}'");
-                Assert.That(protease.Name, Is.EqualTo(expectedNewName), 
-                    $"GetProtease returned wrong protease for '{oldName}'");
-
-                // Verify TryGetProtease works with old name
-                bool found = ProteaseDictionary.TryGetProtease(oldName, out var protease2);
-                Assert.That(found, Is.True, $"TryGetProtease failed for '{oldName}'");
-                Assert.That(protease2.Name, Is.EqualTo(expectedNewName));
-            }
-        }
-
-        /// <summary>
-        /// Tests that GetProtease still works with exact new-style names.
-        /// </summary>
-        [Test]
-        public static void GetProtease_NewStyleName_WorksDirectly()
-        {
-            ProteaseDictionary.ResetToDefaults();
-
-            var protease = ProteaseDictionary.GetProtease("trypsin|P");
-            Assert.That(protease, Is.Not.Null);
-            Assert.That(protease.Name, Is.EqualTo("trypsin|P"));
-
-            bool found = ProteaseDictionary.TryGetProtease("chymotrypsin|P", out var protease2);
-            Assert.That(found, Is.True);
-            Assert.That(protease2.Name, Is.EqualTo("chymotrypsin|P"));
-        }
-
-        /// <summary>
-        /// Tests that GetProtease throws appropriate exception for unknown protease.
-        /// </summary>
-        [Test]
-        public static void GetProtease_UnknownProtease_ThrowsKeyNotFoundException()
-        {
-            ProteaseDictionary.ResetToDefaults();
-
-            Assert.Throws<KeyNotFoundException>(() => ProteaseDictionary.GetProtease("nonexistent protease"));
-            
-            bool found = ProteaseDictionary.TryGetProtease("nonexistent protease", out var protease);
-            Assert.That(found, Is.False);
-            Assert.That(protease, Is.Null);
-        }
-
-        /// <summary>
-        /// Tests that NormalizeProteaseName returns the original name when no pattern matches.
-        /// </summary>
-        [Test]
-        public static void NormalizeProteaseName_NoMatch_ReturnsOriginal()
-        {
-            Assert.That(ProteaseDictionary.NormalizeProteaseName("trypsin"), Is.EqualTo("trypsin"));
-            Assert.That(ProteaseDictionary.NormalizeProteaseName("trypsin|P"), Is.EqualTo("trypsin|P"));
-            Assert.That(ProteaseDictionary.NormalizeProteaseName("custom protease"), Is.EqualTo("custom protease"));
-            Assert.That(ProteaseDictionary.NormalizeProteaseName(""), Is.EqualTo(""));
-            Assert.That(ProteaseDictionary.NormalizeProteaseName(null), Is.Null);
-        }
-
-        /// <summary>
-        /// Tests that loading a protease definition with insufficient fields throws an appropriate exception.
-        /// The parser requires at least 3 fields: Name, Motif, and Specificity.
-        /// </summary>
-        [Test]
-        public static void LoadProteaseDictionary_InsufficientFields_ThrowsWithHelpfulMessage()
-        {
-            string testFile = Path.Combine(TestContext.CurrentContext.TestDirectory, "test_insufficient_fields.tsv");
-            string[] lines =
-            {
-                "Name\tMotif\tSpecificity\tPSI-MS Accession\tPSI-MS Name\tCleavage Modification",
-                "ValidProtease\tK|\tfull\t\t\t",
-                "InvalidProtease\tK|"  // Missing Specificity field - only 2 columns
-            };
-            File.WriteAllLines(testFile, lines);
-
-            var exception = Assert.Throws<MzLibException>(() => ProteaseDictionary.LoadProteaseDictionary(testFile));
-
-            Assert.That(exception.Message, Does.Contain("has only 2 field(s)"));
-            Assert.That(exception.Message, Does.Contain("extend to column 3"));
-            Assert.That(exception.Message, Does.Contain("InvalidProtease"));
-
-            File.Delete(testFile);
-        }
-        /// <summary>
-        /// Tests that protease definitions with only required fields (Name, Motif, Specificity) 
-        /// are parsed correctly, with optional fields defaulting to empty strings.
-        /// </summary>
-        [Test]
-        public static void LoadProteaseDictionary_MinimalFields_DefaultsOptionalFieldsToEmpty()
-        {
-            string testFile = Path.Combine(TestContext.CurrentContext.TestDirectory, "test_minimal_fields.tsv");
-            string[] lines =
-            {
-                "Name\tMotif\tSpecificity\tPSI-MS Accession\tPSI-MS Name\tCleavage Modification",
-                "MinimalProtease\tK|\tfull"  // Only 3 required fields, no optional fields
-            };
-            File.WriteAllLines(testFile, lines);
-
-            var dictionary = ProteaseDictionary.LoadProteaseDictionary(testFile);
-
-            Assert.That(dictionary.ContainsKey("MinimalProtease"), Is.True);
-            var protease = dictionary["MinimalProtease"];
-            Assert.That(protease.Name, Is.EqualTo("MinimalProtease"));
-            Assert.That(protease.CleavageSpecificity, Is.EqualTo(CleavageSpecificity.Full));
-            Assert.That(protease.PsiMsAccessionNumber, Is.EqualTo(string.Empty));
-            Assert.That(protease.PsiMsName, Is.EqualTo(string.Empty));
-            Assert.That(protease.CleavageMod, Is.Null);
-
-            File.Delete(testFile);
-        }
         [Test]
         public static void TestGoodPeptide()
         {
@@ -328,7 +154,7 @@ namespace Test
         public static void TestBadPeptide()
         {
             var prot = new Protein("MNNNKQQXQ", null);
-            var motifList = DigestionMotif.ParseDigestionMotifsFromString("K|");            
+            var motifList = DigestionMotif.ParseDigestionMotifsFromString("K|");
             var protease = new Protease("Custom Protease7", CleavageSpecificity.Full, null, null, motifList);
             ProteaseDictionary.Dictionary.Add(protease.Name, protease);
             DigestionParams digestionParams = new DigestionParams(
@@ -403,7 +229,7 @@ namespace Test
             var ok = prot.Digest(digestionParams, fixedMods, new List<Modification>()).ToList();
 
             Assert.AreEqual(1, ok.Count);
-            
+
             string expectedFullSequence = "[:ProtNmod on M]M[:resMod on M]-[:ProtCmod on M]";
             Assert.AreEqual(expectedFullSequence, ok.First().FullSequence);
             var mods = ok.First().AllModsOneIsNterminus;
@@ -437,7 +263,7 @@ namespace Test
             string expectedSequenceWithChemicalFormulas = "[H]M[H]-[H]";
             double expectedMonoisotopicMass = 5 * GetElement("H").PrincipalIsotope.AtomicMass + Residue.ResidueMonoisotopicMass['M'] + GetElement("O").PrincipalIsotope.AtomicMass;
 
-            
+
 
             // randomly scramble all mods, digest, and ensure the answer is correct. 
             for (int i = 0; i < 10; i++)
@@ -627,106 +453,6 @@ namespace Test
         }
 
         [Test]
-        [TestCase("cRAP_databaseGPTMD.xml")]
-        [TestCase("uniprot_aifm1.fasta")]
-        public static void TestDecoyScramblingIsReproducible(string fileName)
-        {
-            // Load in proteins
-            var dbPath = Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", fileName);
-            DecoyType decoyType = DecoyType.Reverse;
-            List<Protein> proteins1 = null;
-            List<Protein> proteins2 = null;
-            if (fileName.Contains(".xml"))
-            {
-                proteins1 = ProteinDbLoader.LoadProteinXML(dbPath, true, decoyType, null, false, null, out var unknownModifications);
-                proteins2 = ProteinDbLoader.LoadProteinXML(dbPath, true, decoyType, null, false, null, out unknownModifications);
-            }
-            else if (fileName.Contains(".fasta"))
-            {
-                proteins1 = ProteinDbLoader.LoadProteinFasta(dbPath, true, decoyType, false, out var unknownModifications);
-                proteins2 = ProteinDbLoader.LoadProteinFasta(dbPath, true, decoyType, false, out unknownModifications);
-            }
-            else
-            {
-                NUnit.Framework.Assert.Fail("Unknown file type");
-            }
-
-            DigestionParams d = new DigestionParams(
-                        maxMissedCleavages: 1,
-                        minPeptideLength: 5,
-                        initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-            // Digest target proteins
-            var pepsToReplace = proteins1.Where(p => !p.IsDecoy)
-                .SelectMany(p => p.Digest(d, new List<Modification>(), new List<Modification>()).ToList())
-                .Select(pep => pep.BaseSequence)
-                .ToHashSet();
-
-            // Ensure at least one decoy peptide from each protein is problematic and must be replaced
-            var singleDecoyPeptides = proteins1
-                .Where(p => p.IsDecoy)
-                .Select(p => p.Digest(d, new List<Modification>(), new List<Modification>()).Skip(2).Take(1))
-                .Select(pwsm => pwsm.First().BaseSequence)
-                .ToHashSet();
-
-            //modify targetpeptides in place
-            pepsToReplace.UnionWith(singleDecoyPeptides);
-
-            // Scramble every decoy from db1
-            List<Protein> decoys1 = new();
-            foreach (var protein in proteins1.Where(p => p.IsDecoy))
-            {
-                decoys1.Add(DecoySequenceValidator.ScrambleDecoyBioPolymer(protein, d, pepsToReplace));
-            }
-            // Scramble every decoy from db2
-            List<Protein> decoys2 = new();
-            foreach (var protein in proteins2.Where(p => p.IsDecoy))
-            {
-                decoys2.Add(DecoySequenceValidator.ScrambleDecoyBioPolymer(protein, d, pepsToReplace));
-            }
-
-            // check are equivalent lists of proteins
-            Assert.AreEqual(decoys1.Count, decoys2.Count);
-            foreach (var decoyPair in decoys1.Concat(decoys2).GroupBy(p => p.Accession))
-            {
-                Assert.AreEqual(2, decoyPair.Count());
-                Assert.AreEqual(decoyPair.First().BaseSequence, decoyPair.Last().BaseSequence);
-            }
-        }
-
-        [Test]
-        public static void TestDecoyScramblerReplacesPeptides()
-        {
-            DigestionParams d = new DigestionParams(
-                        maxMissedCleavages: 1,
-                        minPeptideLength: 5,
-                        initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            Protein target = new Protein("MEDEEKFVGYKYGVFK", "target");
-            Protein decoy = new Protein("EEDEMKYGVFKFVGYK", "decoy");
-
-            var targetPep = target.Digest(d, new List<Modification>(), new List<Modification>());
-            var decoyPep = decoy.Digest(d, new List<Modification>(), new List<Modification>());
-
-            HashSet<string> targetPepSeqs = targetPep.Select(p => p.FullSequence).ToHashSet();
-            var offendingDecoys = decoyPep.Where(p => targetPepSeqs.Contains(p.FullSequence)).Select(d => d.FullSequence).ToList();
-
-            Assert.AreEqual(2, offendingDecoys.Count);
-
-            Protein scrambledDecoy = DecoySequenceValidator.ScrambleDecoyBioPolymer(decoy,  d, targetPepSeqs, offendingDecoys);
-            var scrambledPep = scrambledDecoy.Digest(d, new List<Modification>(), new List<Modification>());
-
-            Assert.AreEqual(decoyPep.Count(), scrambledPep.Count());
-            Assert.IsFalse(scrambledPep.Any(p => offendingDecoys.Contains(p.FullSequence)));
-
-            // Check to make sure that decoy generation also works in no offending sequences are passed in
-            scrambledDecoy = DecoySequenceValidator.ScrambleDecoyBioPolymer(decoy, d, targetPepSeqs);
-            scrambledPep = scrambledDecoy.Digest(d, new List<Modification>(), new List<Modification>());
-
-            Assert.AreEqual(decoyPep.Count(), scrambledPep.Count());
-            Assert.IsFalse(scrambledPep.Any(p => offendingDecoys.Contains(p.FullSequence)));
-        }
-
-        [Test]
         /// <summary>
         /// Tests that a PeptideWithSetModifications object can be parsed correctly from a string, with mod info
         /// </summary>
@@ -781,145 +507,6 @@ namespace Test
         }
 
         [Test]
-        public static void TestDigestionParamsClone()
-        {
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "top-down",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain,
-                keepNGlycopeptide: true,
-                keepOGlycopeptide: true);
-
-            DigestionParams digestionParamsClone = (DigestionParams)digestionParams.Clone();
-            Assert.AreEqual(digestionParams, digestionParamsClone);
-            Assert.AreEqual(digestionParams.InitiatorMethionineBehavior, digestionParamsClone.InitiatorMethionineBehavior);
-            Assert.AreEqual(digestionParams.MaxMissedCleavages, digestionParamsClone.MaxMissedCleavages);
-            Assert.AreEqual(digestionParams.MaxModificationIsoforms, digestionParamsClone.MaxModificationIsoforms);
-            Assert.AreEqual(digestionParams.MinLength, digestionParamsClone.MinLength);
-            Assert.AreEqual(digestionParams.MaxLength, digestionParamsClone.MaxLength);
-            Assert.AreEqual(digestionParams.MaxMods, digestionParamsClone.MaxMods);
-            Assert.AreEqual(digestionParams.Protease, digestionParamsClone.Protease);
-            Assert.AreEqual(digestionParams.SearchModeType, digestionParamsClone.SearchModeType);
-            Assert.AreEqual(digestionParams.FragmentationTerminus, digestionParamsClone.FragmentationTerminus);
-            Assert.AreEqual(digestionParams.GeneratehUnlabeledProteinsForSilac, digestionParamsClone.GeneratehUnlabeledProteinsForSilac);
-            Assert.AreEqual(digestionParams.KeepNGlycopeptide, digestionParamsClone.KeepNGlycopeptide);
-            Assert.AreEqual(digestionParams.KeepOGlycopeptide, digestionParamsClone.KeepOGlycopeptide);
-            Assert.AreEqual(digestionParams.SpecificProtease, digestionParamsClone.SpecificProtease);
-            NUnit.Framework.Assert.That(!ReferenceEquals(digestionParams, digestionParamsClone));
-
-            digestionParams = new DigestionParams(
-                protease: "top-down",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain,
-                keepNGlycopeptide: true,
-                keepOGlycopeptide: true,
-                maxModificationIsoforms: 5,
-                maxModsForPeptides: 6,
-                maxPeptideLength: 7,
-                searchModeType: CleavageSpecificity.None,
-                fragmentationTerminus: FragmentationTerminus.C,
-                generateUnlabeledProteinsForSilac: false);
-
-            digestionParamsClone = (DigestionParams)digestionParams.Clone();
-            Assert.AreEqual(digestionParams, digestionParamsClone);
-            Assert.AreEqual(digestionParams.InitiatorMethionineBehavior, digestionParamsClone.InitiatorMethionineBehavior);
-            Assert.AreEqual(digestionParams.MaxMissedCleavages, digestionParamsClone.MaxMissedCleavages);
-            Assert.AreEqual(digestionParams.MaxModificationIsoforms, digestionParamsClone.MaxModificationIsoforms);
-            Assert.AreEqual(digestionParams.MinLength, digestionParamsClone.MinLength);
-            Assert.AreEqual(digestionParams.MaxLength, digestionParamsClone.MaxLength);
-            Assert.AreEqual(digestionParams.MaxMods, digestionParamsClone.MaxMods);
-            Assert.AreEqual(digestionParams.Protease, digestionParamsClone.Protease);
-            Assert.AreEqual(digestionParams.SearchModeType, digestionParamsClone.SearchModeType);
-            Assert.AreEqual(digestionParams.FragmentationTerminus, digestionParamsClone.FragmentationTerminus);
-            Assert.AreEqual(digestionParams.GeneratehUnlabeledProteinsForSilac, digestionParamsClone.GeneratehUnlabeledProteinsForSilac);
-            Assert.AreEqual(digestionParams.KeepNGlycopeptide, digestionParamsClone.KeepNGlycopeptide);
-            Assert.AreEqual(digestionParams.KeepOGlycopeptide, digestionParamsClone.KeepOGlycopeptide);
-            Assert.AreEqual(digestionParams.SpecificProtease, digestionParamsClone.SpecificProtease);
-            NUnit.Framework.Assert.That(!ReferenceEquals(digestionParams, digestionParamsClone));
-        }
-
-
-
-        [Test]
-        public static void TestDigestionParamsCloneWithNewTerminus()
-        {
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "top-down",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain,
-                keepNGlycopeptide: true,
-                keepOGlycopeptide: true);
-
-            DigestionParams digestionParamsClone = (DigestionParams)digestionParams.Clone(FragmentationTerminus.N);
-            Assert.AreNotEqual(digestionParams, digestionParamsClone);
-            Assert.AreEqual(digestionParams.InitiatorMethionineBehavior, digestionParamsClone.InitiatorMethionineBehavior);
-            Assert.AreEqual(digestionParams.MaxMissedCleavages, digestionParamsClone.MaxMissedCleavages);
-            Assert.AreEqual(digestionParams.MaxModificationIsoforms, digestionParamsClone.MaxModificationIsoforms);
-            Assert.AreEqual(digestionParams.MinLength, digestionParamsClone.MinLength);
-            Assert.AreEqual(digestionParams.MaxLength, digestionParamsClone.MaxLength);
-            Assert.AreEqual(digestionParams.MaxMods, digestionParamsClone.MaxMods);
-            Assert.AreEqual(digestionParams.Protease, digestionParamsClone.Protease);
-            Assert.AreEqual(digestionParams.SearchModeType, digestionParamsClone.SearchModeType);
-            Assert.AreEqual(FragmentationTerminus.N, digestionParamsClone.FragmentationTerminus);
-            Assert.AreEqual(digestionParams.GeneratehUnlabeledProteinsForSilac, digestionParamsClone.GeneratehUnlabeledProteinsForSilac);
-            Assert.AreEqual(digestionParams.KeepNGlycopeptide, digestionParamsClone.KeepNGlycopeptide);
-            Assert.AreEqual(digestionParams.KeepOGlycopeptide, digestionParamsClone.KeepOGlycopeptide);
-            Assert.AreEqual(digestionParams.SpecificProtease, digestionParamsClone.SpecificProtease);
-            NUnit.Framework.Assert.That(!ReferenceEquals(digestionParams, digestionParamsClone));
-
-            digestionParams = new DigestionParams(
-                protease: "top-down",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain,
-                keepNGlycopeptide: true,
-                keepOGlycopeptide: true,
-                maxModificationIsoforms: 5,
-                maxModsForPeptides: 6,
-                maxPeptideLength: 7,
-                searchModeType: CleavageSpecificity.None,
-                fragmentationTerminus: FragmentationTerminus.None,
-                generateUnlabeledProteinsForSilac: false);
-
-            digestionParamsClone = (DigestionParams)digestionParams.Clone(FragmentationTerminus.N);
-            Assert.AreNotEqual(digestionParams, digestionParamsClone);
-            Assert.AreEqual(digestionParams.InitiatorMethionineBehavior, digestionParamsClone.InitiatorMethionineBehavior);
-            Assert.AreEqual(digestionParams.MaxMissedCleavages, digestionParamsClone.MaxMissedCleavages);
-            Assert.AreEqual(digestionParams.MaxModificationIsoforms, digestionParamsClone.MaxModificationIsoforms);
-            Assert.AreEqual(digestionParams.MinLength, digestionParamsClone.MinLength);
-            Assert.AreEqual(digestionParams.MaxLength, digestionParamsClone.MaxLength);
-            Assert.AreEqual(digestionParams.MaxMods, digestionParamsClone.MaxMods);
-            Assert.AreEqual(ProteaseDictionary.Dictionary["singleN"], digestionParamsClone.Protease);
-            Assert.AreEqual(digestionParams.SearchModeType, digestionParamsClone.SearchModeType);
-            Assert.AreEqual(FragmentationTerminus.N, digestionParamsClone.FragmentationTerminus);
-            Assert.AreEqual(digestionParams.GeneratehUnlabeledProteinsForSilac, digestionParamsClone.GeneratehUnlabeledProteinsForSilac);
-            Assert.AreEqual(digestionParams.KeepNGlycopeptide, digestionParamsClone.KeepNGlycopeptide);
-            Assert.AreEqual(digestionParams.KeepOGlycopeptide, digestionParamsClone.KeepOGlycopeptide);
-            Assert.AreEqual(digestionParams.SpecificProtease, digestionParamsClone.SpecificProtease);
-            NUnit.Framework.Assert.That(!ReferenceEquals(digestionParams, digestionParamsClone));
-
-            digestionParamsClone = (DigestionParams)digestionParams.Clone(FragmentationTerminus.C);
-            Assert.AreNotEqual(digestionParams, digestionParamsClone);
-            Assert.AreEqual(digestionParams.InitiatorMethionineBehavior, digestionParamsClone.InitiatorMethionineBehavior);
-            Assert.AreEqual(digestionParams.MaxMissedCleavages, digestionParamsClone.MaxMissedCleavages);
-            Assert.AreEqual(digestionParams.MaxModificationIsoforms, digestionParamsClone.MaxModificationIsoforms);
-            Assert.AreEqual(digestionParams.MinLength, digestionParamsClone.MinLength);
-            Assert.AreEqual(digestionParams.MaxLength, digestionParamsClone.MaxLength);
-            Assert.AreEqual(digestionParams.MaxMods, digestionParamsClone.MaxMods);
-            Assert.AreEqual(ProteaseDictionary.Dictionary["singleC"], digestionParamsClone.Protease);
-            Assert.AreEqual(digestionParams.SearchModeType, digestionParamsClone.SearchModeType);
-            Assert.AreEqual(FragmentationTerminus.C, digestionParamsClone.FragmentationTerminus);
-            Assert.AreEqual(digestionParams.GeneratehUnlabeledProteinsForSilac, digestionParamsClone.GeneratehUnlabeledProteinsForSilac);
-            Assert.AreEqual(digestionParams.KeepNGlycopeptide, digestionParamsClone.KeepNGlycopeptide);
-            Assert.AreEqual(digestionParams.KeepOGlycopeptide, digestionParamsClone.KeepOGlycopeptide);
-            Assert.AreEqual(digestionParams.SpecificProtease, digestionParamsClone.SpecificProtease);
-            NUnit.Framework.Assert.That(!ReferenceEquals(digestionParams, digestionParamsClone));
-        }
-
-        [Test]
         public static void TestWhenFixedModIsSamePositionAsUniProtModWithDigestion()
         {
             var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
@@ -958,114 +545,6 @@ namespace Test
             };
 
             CollectionAssert.AreEquivalent(expectedFullSequences, fullSequences);
-        }
-        /// <summary>
-        /// Tests the custom protease dictionary functionality including:
-        /// - Loading custom proteases from a file and merging into the main dictionary
-        /// - Overwriting existing proteases with custom definitions
-        /// - Adding new proteases not in the default set
-        /// - Using custom proteases for protein digestion
-        /// - Resetting to default proteases
-        /// 
-        /// Custom protease files use TSV format with columns:
-        /// Name, Motif, Specificity, PSI-MS Accession, PSI-MS Name, Cleavage Modification
-        /// 
-        /// Merge rules:
-        /// - If protease name matches existing entry: OVERWRITES the built-in definition
-        /// - If protease name is new: ADDS to the dictionary
-        /// </summary>
-        [Test]
-        public static void LoadAndMergeCustomProteases_OverwritesAndAddsProteases()
-        {
-            // Arrange - capture initial state
-            int initialProteaseCount = ProteaseDictionary.Dictionary.Count;
-            var originalTrypsin = ProteaseDictionary.Dictionary["trypsin|P"];
-            Assert.That(originalTrypsin.DigestionMotifs.Count, Is.EqualTo(2)); // K[P]| and R[P]|
-            // Verify original trypsin|P cleaves after K and R (not before P)
-            Assert.That(originalTrypsin.DigestionMotifs.Any(m => m.InducingCleavage == "K"), Is.True);
-            Assert.That(originalTrypsin.DigestionMotifs.Any(m => m.InducingCleavage == "R"), Is.True);
-
-            // Create a custom protease file that:
-            // 1. Overrides trypsin|P with a completely different (nonsense) cleavage rule: cleave after L unless followed by P
-            // 2. Adds a completely new custom protease
-            string customProteaseFile = Path.Combine(TestContext.CurrentContext.TestDirectory, "test_custom_proteases.tsv");
-            string[] lines =
-            {
-                "Name\tMotif\tSpecificity\tPSI-MS Accession\tPSI-MS Name\tCleavage Modification",
-                "trypsin|P\tL[P]|\tfull\tMS:1001313\tTrypsin\t",  // Override: change from K[P]|,R[P]| to L[P]| (nonsense rule for testing)
-                "MyLabProtease\tE|\tfull\t\tCustom Glu-C variant\t"  // New: cleaves after glutamate
-            };
-            File.WriteAllLines(customProteaseFile, lines);
-
-            try
-            {
-                // Act - merge custom proteases
-                var addedOrUpdated = ProteaseDictionary.LoadAndMergeCustomProteases(customProteaseFile);
-
-                // Assert - verify merge results
-                Assert.That(addedOrUpdated.Count, Is.EqualTo(2));
-                Assert.That(addedOrUpdated, Contains.Item("trypsin|P"));
-                Assert.That(addedOrUpdated, Contains.Item("MyLabProtease"));
-                Assert.That(ProteaseDictionary.Dictionary.Count, Is.EqualTo(initialProteaseCount + 1)); // Only one new protease added (MyLabProtease); trypsin|P was overwritten
-
-                // Verify trypsin|P was overwritten with the new L[P]| motif
-                var customTrypsin = ProteaseDictionary.Dictionary["trypsin|P"];
-                Assert.That(customTrypsin.DigestionMotifs.Count, Is.EqualTo(1)); // Now only L[P]|
-                Assert.That(customTrypsin.DigestionMotifs[0].InducingCleavage, Is.EqualTo("L"));
-                Assert.That(customTrypsin.DigestionMotifs[0].PreventingCleavage, Is.EqualTo("P"));
-
-                // Verify new protease was added
-                Assert.That(ProteaseDictionary.Dictionary.ContainsKey("MyLabProtease"), Is.True);
-                var myLabProtease = ProteaseDictionary.Dictionary["MyLabProtease"];
-                Assert.That(myLabProtease.DigestionMotifs.Count, Is.EqualTo(1));
-                Assert.That(myLabProtease.DigestionMotifs[0].InducingCleavage, Is.EqualTo("E"));
-
-                // Verify custom proteases work for digestion
-                // Protein with L's for testing the overwritten trypsin|P
-                // Note: L at position 8 is followed by 'E' (not P), so cleavage will occur there
-                var protein = new Protein("PEPTIDELEPEPTIDER", null);
-
-                // Custom trypsin|P should now cleave after L (unless followed by P)
-                var customTrypsinParams = new DigestionParams(
-                    protease: "trypsin|P",
-                    maxMissedCleavages: 0,
-                    minPeptideLength: 1);
-                var customDigest = protein.Digest(customTrypsinParams, new List<Modification>(), new List<Modification>()).ToList();
-                // Should cleave after L at position 8 (L is followed by E, not P), producing: PEPTIDEL, EPEPTIDER
-                Assert.That(customDigest.Count, Is.EqualTo(2));
-                Assert.That(customDigest[0].BaseSequence, Is.EqualTo("PEPTIDEL"));
-                Assert.That(customDigest[1].BaseSequence, Is.EqualTo("EPEPTIDER"));
-
-                // New protease should work
-                var myLabParams = new DigestionParams(
-                    protease: "MyLabProtease",
-                    maxMissedCleavages: 0,
-                    minPeptideLength: 1);
-                var myLabDigest = protein.Digest(myLabParams, new List<Modification>(), new List<Modification>()).ToList();
-                Assert.That(myLabDigest.Count, Is.EqualTo(6)); // PE, PTIDE, LE, PE, PTIDE, R (cleaves after each E)
-
-                // Act - reset to defaults
-                ProteaseDictionary.ResetToDefaults();
-
-                // Assert - verify reset worked
-                Assert.That(ProteaseDictionary.Dictionary.Count, Is.EqualTo(initialProteaseCount));
-                Assert.That(ProteaseDictionary.Dictionary.ContainsKey("MyLabProtease"), Is.False);
-
-                // Verify trypsin|P is back to original behavior (K[P]| and R[P]|)
-                var restoredTrypsin = ProteaseDictionary.Dictionary["trypsin|P"];
-                Assert.That(restoredTrypsin.DigestionMotifs.Count, Is.EqualTo(2));
-                Assert.That(restoredTrypsin.DigestionMotifs.Any(m => m.InducingCleavage == "K"), Is.True);
-                Assert.That(restoredTrypsin.DigestionMotifs.Any(m => m.InducingCleavage == "R"), Is.True);
-            }
-            finally
-            {
-                // Cleanup - ensure dictionary is reset even if test fails
-                ProteaseDictionary.ResetToDefaults();
-                if (File.Exists(customProteaseFile))
-                {
-                    File.Delete(customProteaseFile);
-                }
-            }
         }
     }
 }
